refactor: remove unnecessary plugin dependency from root package.json
The cypress-ncatestify-plugin is not needed in the root package.json
since this is just an npm initialization template. The plugin is only
required in the template directory that users receive when they
initialize a new project.

This cleanup reduces confusion about which package.json to update and
clarifies that only template/package.json needs maintenance.